requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.115.14
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.20.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.10.0
joblib==1.5.1
mysql-connector-python==9.4.0
nltk==3.9.1
openai==1.91.0
outcome==1.3.0.post0
pandas==2.3.0
pika==1.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.6
pydantic_core==2.33.2
PyMySQL==1.1.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
schedule==1.2.2
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
starlette==0.46.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.35.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
